+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version X.Y.Z (YYYYMMDDXX)
+
+- Add Moodle plugin CI for all supported Moodle versions
+
+
 ## Version 0.0.1 (2025081900)
 
 - Initial stub release