+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version X.Y.Z (YYYYMMDDXX)
 
 - Implement store, retrieve, and delete functionality using the Moodle Filestore backend.
+- Check free space in the moodledata directory to determine storage availability.
 - Add Moodle plugin CI for all supported Moodle versions