+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Implement store, retrieve, and delete functionality using the Moodle Filestore backend
 - Check free space in the moodledata directory to determine storage availability
+- Add privacy provider class
 - Create unit tests
 - Add Moodle plugin CI for all supported Moodle versions